Fix linked list allocator alignment bug
The linked list allocator searched for a block that was big enough to
accomodated the requested size even when aligned, and also the
difference between the block's address and the aligned address was big
enough to hold a new block.

However this lead to a situation where the available bocks were big
enough to accomodate the requested size, but because the aligned address
was very close to the starting address, the block could not be used.

To fix this, keep increasing the aligned address untill the diference is
big enough. Then we can normally check if this new aligned address
exceeds the boudaries of the block or not, as usual.