Add debug utilities in cooperative::task module
Add some basic debug formating in the Task struct.

Add a method from creating task id's from usize. This is an unsafe
method since in normal operation TaskId's must be unique, so when a user
creates one directly from a usize, they must make sure that it is
unique. This function should not be used unless debuging.