Add breakpoint dummy function
Add a function named 'breakpoint' that does nothing. This is useful when
debugging with gdb and one wants to add breakpoints directly from the
editor. To do this, simply make a call to this function at the point
where you want the breakpoint, and then, in gdb, add a breakpoint
whenever this function is called.

This is probably a bad hack, but it is the only way I know how to do
this. For some reason the core::intrinsics::breakpoint function doesn't
work.