Fix bug in linked_list_allocator module
Due to how I had originally thought linked_listed_allocator can be
implemented, when a memory region was deallocated, the block that was
added in the linked list had a size that was the length of the memory
region plus the alignment.

However the implementation was redesigned. But the old way of
calculating the block's size remained, even though it was now wrong.

Now the allocator correctly sets the size of the new block to be that of
the deallocated memory region.