Add a Spawner to spawn tasks after executor starts
When the `run()` method is called on the executor, there was no way of
spawning new tasks. The new Spawner struct fixes this by having a queue
that is shared with the executor, and pushing newly spawned tasks there.

Before calling `run()` on the executor, one can get a spawner by calling
`get_spawner()`. This new spawner can be cloned and passed to new tasks,
which in turn can use it to spawn tasks on their own.