x86: add io_wait function
Sometimes, a piece of code needs to create an artificial delay before
proceeding to the next instruction. Normally we would rely on some kind
of interrupt in order to no hold the cpu with no reason. However some
times an interrupt is not available, and we have to wait.

To achieve this we can read from an unused port, which will wait untill
the port is written. This should be independent of the CPU speed.

Idea taken from: https://wiki.osdev.org/Inline_Assembly/Examples#IO_WAIT